fix(projects/products/stable/platform_ide/backend): enforce strict test layout and main entrypoint contract
Part of #1075
