+++ projects/products/stable/platform_ide/backend/src/app/tests/mod.rs
@@ -1,0 +1,2 @@
+//! projects/products/stable/platform_ide/backend/src/app/tests/mod.rs
+mod ide_config;
