+++ projects/products/stable/platform_ide/backend/src/auth/mod.rs
@@ -1,4 +1,7 @@
-// projects/products/stable/platform_ide/backend/src/auth/mod.rs
+//! projects/products/stable/platform_ide/backend/src/auth/mod.rs
 pub mod session;
 
 pub use session::Session;
+
+#[cfg(test)]
+mod tests;
