+++ projects/products/stable/platform_ide/backend/src/changes/tests/mod.rs
@@ -1,0 +1,3 @@
+//! projects/products/stable/platform_ide/backend/src/changes/tests/mod.rs
+mod change_set;
+mod patch_entry;
