+++ projects/products/stable/platform_ide/backend/src/client/mod.rs
@@ -1,4 +1,7 @@
-// projects/products/stable/platform_ide/backend/src/client/mod.rs
+//! projects/products/stable/platform_ide/backend/src/client/mod.rs
 pub mod platform_client;
 
 pub use platform_client::PlatformClient;
+
+#[cfg(test)]
+mod tests;
