+++ projects/products/stable/platform_ide/backend/src/client/tests/mod.rs
@@ -1,0 +1,2 @@
+//! projects/products/stable/platform_ide/backend/src/client/tests/mod.rs
+mod platform_client;
