+++ projects/products/stable/platform_ide/backend/src/diff/local_diff.rs
@@ -1,4 +1,4 @@
-// projects/products/stable/platform_ide/backend/src/diff/local_diff.rs
+//! projects/products/stable/platform_ide/backend/src/diff/local_diff.rs
 use crate::diff::diff_line::DiffLine;
 use crate::editor::FileBuffer;
 use crate::slices::AllowedPath;
@@ -96,49 +96,5 @@ fn backtrack(
     } else {
         backtrack(dp, a, b, i, j - 1, out);
         out.push(DiffLine::Added(b[j - 1].to_string()));
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::editor::FileBuffer;
-    use crate::slices::AllowedPath;
-
-    fn allowed(path: &str) -> AllowedPath {
-        AllowedPath::new_validated(path.to_string())
-    }
-
-    #[test]
-    fn no_changes_produces_only_context() {
-        let buf = FileBuffer::open(allowed("a.txt"), b"line1\nline2\n".to_vec());
-        let diff = LocalDiff::from_buffer(&buf);
-        assert!(!diff.has_changes());
-        assert!(diff.lines.iter().all(|l| matches!(l, DiffLine::Context(_))));
-    }
-
-    #[test]
-    fn added_line_detected() {
-        let mut buf = FileBuffer::open(allowed("a.txt"), b"line1\n".to_vec());
-        buf.write(b"line1\nline2\n".to_vec());
-        let diff = LocalDiff::from_buffer(&buf);
-        assert!(diff.has_changes());
-        assert!(diff.lines.iter().any(|l| matches!(l, DiffLine::Added(_))));
-    }
-
-    #[test]
-    fn removed_line_detected() {
-        let mut buf = FileBuffer::open(allowed("a.txt"), b"line1\nline2\n".to_vec());
-        buf.write(b"line1\n".to_vec());
-        let diff = LocalDiff::from_buffer(&buf);
-        assert!(diff.has_changes());
-        assert!(diff.lines.iter().any(|l| matches!(l, DiffLine::Removed(_))));
-    }
-
-    #[test]
-    fn path_is_preserved() {
-        let buf = FileBuffer::open(allowed("src/lib.rs"), b"fn foo() {}".to_vec());
-        let diff = LocalDiff::from_buffer(&buf);
-        assert_eq!(diff.path.as_str(), "src/lib.rs");
     }
 }
