+++ projects/products/stable/platform_ide/backend/src/diff/tests/mod.rs
@@ -1,0 +1,2 @@
+//! projects/products/stable/platform_ide/backend/src/diff/tests/mod.rs
+mod local_diff;
