+++ projects/products/stable/platform_ide/backend/src/editor/mod.rs
@@ -1,4 +1,7 @@
-// projects/products/stable/platform_ide/backend/src/editor/mod.rs
+//! projects/products/stable/platform_ide/backend/src/editor/mod.rs
 pub mod file_buffer;
 
 pub use file_buffer::FileBuffer;
+
+#[cfg(test)]
+mod tests;
