+++ projects/products/stable/platform_ide/backend/src/editor/tests/mod.rs
@@ -1,0 +1,2 @@
+//! projects/products/stable/platform_ide/backend/src/editor/tests/mod.rs
+mod file_buffer;
