+++ projects/products/stable/platform_ide/backend/src/offline/mod.rs
@@ -1,4 +1,7 @@
-// projects/products/stable/platform_ide/backend/src/offline/mod.rs
+//! projects/products/stable/platform_ide/backend/src/offline/mod.rs
 pub mod offline_policy;
 
 pub use offline_policy::OfflinePolicy;
+
+#[cfg(test)]
+mod tests;
