+++ projects/products/stable/platform_ide/backend/src/offline/offline_policy.rs
@@ -1,4 +1,4 @@
-// projects/products/stable/platform_ide/backend/src/offline/offline_policy.rs
+//! projects/products/stable/platform_ide/backend/src/offline/offline_policy.rs
 use serde::{Deserialize, Serialize};
 
 use crate::errors::IdeError;
@@ -37,33 +37,5 @@ impl OfflinePolicy {
         } else {
             Err(IdeError::OfflineNotPermitted)
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn default_is_disabled() {
-        let policy = OfflinePolicy::default();
-        assert!(!policy.is_allowed());
-        assert!(policy.require_allowed().is_err());
-    }
-
-    #[test]
-    fn disabled_helper_matches_default() {
-        let policy = OfflinePolicy::disabled();
-        assert!(!policy.is_allowed());
-    }
-
-    #[test]
-    fn allowed_policy_permits() {
-        let policy = OfflinePolicy {
-            allowed: true,
-            notice: Some("Admin has approved offline access.".to_string()),
-        };
-        assert!(policy.is_allowed());
-        assert!(policy.require_allowed().is_ok());
     }
 }
