+++ projects/products/stable/platform_ide/backend/src/offline/tests/mod.rs
@@ -1,0 +1,2 @@
+//! projects/products/stable/platform_ide/backend/src/offline/tests/mod.rs
+mod offline_policy;
