+++ projects/products/stable/platform_ide/backend/src/slices/allowed_path.rs
@@ -1,4 +1,4 @@
-// projects/products/stable/platform_ide/backend/src/slices/allowed_path.rs
+//! projects/products/stable/platform_ide/backend/src/slices/allowed_path.rs
 use serde::{Deserialize, Serialize};
 
 /// A validated file path that has been confirmed to exist within the slice
@@ -54,38 +54,4 @@ pub(super) fn is_safe_path(raw: &str) -> bool {
         }
     }
     true
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn safe_path_accepts_normal_paths() {
-        assert!(is_safe_path("src/main.rs"));
-        assert!(is_safe_path("README.md"));
-        assert!(is_safe_path("a/b/c/d.txt"));
-    }
-
-    #[test]
-    fn safe_path_rejects_traversal() {
-        assert!(!is_safe_path("../etc/passwd"));
-        assert!(!is_safe_path("src/../../secret"));
-        assert!(!is_safe_path("/absolute/path"));
-        assert!(!is_safe_path(""));
-    }
-
-    #[test]
-    fn allowed_path_display() {
-        let p = AllowedPath::new_validated("src/lib.rs".to_string());
-        assert_eq!(p.to_string(), "src/lib.rs");
-        assert_eq!(p.as_str(), "src/lib.rs");
-    }
-
-    #[test]
-    fn allowed_path_equality() {
-        let a = AllowedPath::new_validated("src/lib.rs".to_string());
-        let b = AllowedPath::new_validated("src/lib.rs".to_string());
-        assert_eq!(a, b);
-    }
 }
