+++ projects/products/stable/platform_ide/backend/src/slices/tests/mod.rs
@@ -1,0 +1,3 @@
+//! projects/products/stable/platform_ide/backend/src/slices/tests/mod.rs
+mod allowed_path;
+mod slice_manifest;
