+++ projects/products/stable/platform_ide/backend/src/tests/mod.rs
@@ -1,0 +1,2 @@
+//! projects/products/stable/platform_ide/backend/src/tests/mod.rs
+mod app;
