+++ projects/products/stable/platform_ide/backend/src/verification/mod.rs
@@ -1,4 +1,4 @@
-// projects/products/stable/platform_ide/backend/src/verification/mod.rs
+//! projects/products/stable/platform_ide/backend/src/verification/mod.rs
 pub mod finding_severity;
 pub mod raw_finding;
 pub mod result_view;
@@ -8,3 +8,6 @@ pub use finding_severity::FindingSeverity;
 pub use raw_finding::RawFinding;
 pub use result_view::VerificationResultView;
 pub use verification_finding::VerificationFinding;
+
+#[cfg(test)]
+mod tests;
