+++ projects/products/stable/platform_ide/backend/src/verification/result_view.rs
@@ -1,4 +1,4 @@
-// projects/products/stable/platform_ide/backend/src/verification/result_view.rs
+//! projects/products/stable/platform_ide/backend/src/verification/result_view.rs
 use crate::slices::SliceManifest;
 use crate::verification::{FindingSeverity, RawFinding, VerificationFinding};
 use serde::{Deserialize, Serialize};
@@ -75,94 +75,3 @@ impl ResultView {
 }
 
 pub type VerificationResultView = ResultView;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::slices::SliceManifest;
-
-    fn manifest() -> SliceManifest {
-        SliceManifest::new("issue-1", "abc", ["src/main.rs", "README.md"])
-    }
-
-    fn raw(severity: FindingSeverity, summary: &str, path: Option<&str>) -> RawFinding {
-        RawFinding {
-            severity,
-            summary: summary.to_string(),
-            path: path.map(str::to_string),
-            line: path.map(|_| 10),
-        }
-    }
-
-    #[test]
-    fn allowed_path_finding_is_shown() {
-        let view = ResultView::from_raw(
-            false,
-            [raw(
-                FindingSeverity::Error,
-                "syntax error",
-                Some("src/main.rs"),
-            )],
-            &manifest(),
-        );
-        assert_eq!(view.findings.len(), 1);
-        let f = &view.findings[0];
-        assert_eq!(f.path.as_deref(), Some("src/main.rs"));
-        assert!(f.line.is_some());
-        assert_eq!(f.summary, "syntax error");
-    }
-
-    #[test]
-    fn forbidden_path_error_finding_becomes_generic() {
-        let view = ResultView::from_raw(
-            false,
-            [raw(
-                FindingSeverity::Error,
-                "secret error",
-                Some("forbidden/internal.rs"),
-            )],
-            &manifest(),
-        );
-        assert_eq!(view.findings.len(), 1);
-        let f = &view.findings[0];
-        assert!(f.path.is_none(), "forbidden path must not appear in output");
-        assert!(f.line.is_none(), "line must not appear for forbidden path");
-        assert!(!f.summary.contains("forbidden"), "summary leaks path info");
-        assert!(
-            !f.summary.contains("internal.rs"),
-            "summary leaks path info"
-        );
-    }
-
-    #[test]
-    fn forbidden_path_info_finding_is_suppressed() {
-        let view = ResultView::from_raw(
-            true,
-            [raw(
-                FindingSeverity::Info,
-                "style note",
-                Some("forbidden/internal.rs"),
-            )],
-            &manifest(),
-        );
-        assert!(view.findings.is_empty());
-    }
-
-    #[test]
-    fn no_path_finding_is_always_shown() {
-        let view = ResultView::from_raw(
-            false,
-            [raw(FindingSeverity::Warning, "global warning", None)],
-            &manifest(),
-        );
-        assert_eq!(view.findings.len(), 1);
-        assert!(view.findings[0].path.is_none());
-    }
-
-    #[test]
-    fn healthy_result_has_no_findings() {
-        let view = ResultView::healthy();
-        assert!(view.healthy);
-        assert!(view.findings.is_empty());
-    }
-}
