+++ projects/products/stable/platform_ide/backend/src/verification/tests/mod.rs
@@ -1,0 +1,2 @@
+//! projects/products/stable/platform_ide/backend/src/verification/tests/mod.rs
+mod result_view;
